requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pytest==8.1.1
pytest-cov==5.0.0
httpx==0.27.0